netdev-dpdk: add support for packet capture
This patch add hooks to DPDK netdev send/receive functions
to make a copy of the mbuf and push to a rte ring.

Another separated thread will consume the packets and in
the future dump to a file.

To enable packet capture:
$ ovs-appctl netdev-dpdk/set-pcap <netdev> rx|tx <filename>

To disable packet capture:
$ ovs-appctl netdev-dpdk/set-pcap <netdev> rx|tx

Bidirectional traffic:
       No PCAP: 8123635.00  (100%)
 TX or RX only: 5466910.00  (-32%)
     TX and RX: 4079750.00  (-50%)

Unidirectional traffic:
       No PCAP: 14880738.00  (100%)
 TX or RX only:  9878410.00  (-33%)
     TX and RX:  9878410.00  (-33%)

Signed-off-by: Flavio Leitner <fbl@redhat.com>