Enable IP checksum offloading.
The IPv4 and IPv6 flags which identifies the network protocol
are used to indicate IP checksum offload. Add a new flag to
indicate IP checksum offloading.

When encapsulate a packet with that flag, set the checksum
of the inner IP header since that is not yet supported.

Calculate the IP csum when the packet is going to be sent over
a device that doesn't support the feature.

The netdev receiving packets is supposed to provide the flags
indicating if the IP csum was verified and it is OK or BAD,
otherwise the stack will check when appropriate by software.

fbl